+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - 优化 `f-tag` 组件内部逻辑
 - 优化 `f-button` 组件实现
 - 优化 `f-ripple` 组件实现
+- 废除 `f-ripple` 组件 `no-select` 配置项
 
 ## 0.9.0-alpha.7 (2022-12-01)